Cast (size_t) to (unsigned int).
It's not always automatically an unsigned int on other platforms, e.g. Mac OS X.

+++ src/oping.c
@@ -322,7 +322,7 @@ int main (int argc, char **argv)
 		ping_iterator_get_info (iter, PING_INFO_DATA, NULL, &buffer_size);
 
 		printf ("PING %s (%s) %u bytes of data.\n",
-				context->host, context->addr, buffer_size);
+				context->host, context->addr, (unsigned int) buffer_size);
 
 		ping_iterator_set_context (iter, (void *) context);
 	}
